// old src/types/options.d.ts
export interface MessageCommandOptions {
    name: string;
    aliases?: Array<string>;
    separator?: string;
    description?: string;
    info?: string;
    category?: string;
    prefix?: Array<string> | string;
    regex?: string;
    channelType?: Array<number>;
    cooldown?: number;
    ignore?: {
        owner?: boolean;
        users?: Array<string>;
        channels?: Array<string>;
        guilds?: Array<string>;
    };
    strict?: {
        owner?: boolean;
        users?: Array<string>;
        channels?: Array<string>;
        guilds?: Array<string>;
    };
    permissions?: {
        client?: Array<string>;
        user?: Array<string>;
        dm?: boolean;
    };
    args?: Array<MessageArgumentOptions>;
}
// ...
export interface MessageArgumentOptions {
    type: MessageCommandArgsTypes;
    content?: Array<string>;
    match?: string;
    name: string;
    prompt?: {
        reply: string;
        send: string;
    };
    default?: string | Function;
}